feat(account): Primal/Damus-style drawer + per-section screens (#100)
Replaces the single long-scroll AccountScreen with a left-side drawer
that hosts per-section screens (Profile, Wallets, Nostr, On-chain,
Display, Support, About). Drawer header shows the enlarged avatar,
display name, and truncated npub; a Sign Out row sits above a pinned
version footer. Tapping the top-right profile icon in any tab opens
the drawer.

- Adds @react-navigation/drawer and wires AccountDrawer around MainTabs
- New AccountDrawerContent renders header + section rows + sign-out +
  version footer per #100's spec (no follower counts)
- Splits AccountScreen into screens under src/screens/account/ with
  a shared AccountScreenLayout + sharedStyles
- Moves the triple-tap-to-toggle dev-mode affordance to the About
  screen's version string; drawer footer shows the version read-only
- Updates existing Maestro tests to navigate through the drawer and
  use the new sign-out/back selectors; adds tests/e2e/test-account-
  drawer.yaml covering drawer open, per-section navigation, and the
  dev-mode toggle via About

Relays and Privacy+Security drawer entries from the issue spec are
deliberately out of scope for v1 — they need new settings screens
that do not exist today; #100 is explicit that v1 is a pure
restructure.
