+++ package.json
@@ -22,6 +22,7 @@
     "@noble/hashes": "^2.0.1",
     "@react-native-async-storage/async-storage": "2.2.0",
     "@react-navigation/bottom-tabs": "^7.15.8",
+    "@react-navigation/drawer": "^7.9.8",
     "@react-navigation/native": "^7.2.1",
     "@react-navigation/native-stack": "^7.14.9",
     "@scure/btc-signer": "^2.0.1",
