+++ src/screens/FriendsScreen.tsx
@@ -391,7 +391,7 @@ const FriendsScreen: React.FC = () => {
             </Text>
             <TouchableOpacity
               style={styles.connectButton}
-              onPress={() => navigation.navigate('Account')}
+              onPress={() => navigation.getParent()?.dispatch({ type: 'OPEN_DRAWER' })}
             >
               <Text style={styles.connectButtonText}>Go to Account</Text>
             </TouchableOpacity>
