+++ tests/e2e/test-accept-nostr-profile.yaml
@@ -6,7 +6,8 @@ name: Accept Nostr profile (adopt lud16 + displayName)
 # us a lightning address (lud16) without needing to add a wallet first —
 # the Receive → Address flow then works end-to-end.
 
-# Trigger the login flow (which surfaces the profile-update alert)
+# Open the drawer and go to the Display sub-screen (where the lightning
+# address lives after the drawer refactor in #100).
 - tapOn: { id: 'tab-home', optional: true }
 - waitForAnimationToEnd: { timeout: 2000 }
 - tapOn: { id: 'profile-icon' }
@@ -14,6 +15,9 @@ name: Accept Nostr profile (adopt lud16 + displayName)
 
 # If the alert is showing from previous session, adopt the profile
 - tapOn: { text: 'UPDATE', optional: true }
+- waitForAnimationToEnd: { timeout: 2000 }
+
+- tapOn: { id: 'drawer-row-display' }
 - waitForAnimationToEnd: { timeout: 2000 }
 
 # Scroll down to find the lightning address row to confirm it's set
