+++ tests/e2e/test-login-dev.yaml
@@ -11,7 +11,7 @@ name: Nostr Login Test (dev variant)
 - waitForAnimationToEnd:
     timeout: 10000
 
-# Navigate to Account via profile icon on Home
+# Open the account drawer
 - tapOn:
     id: 'tab-home'
     optional: true
@@ -22,20 +22,26 @@ name: Nostr Login Test (dev variant)
 - waitForAnimationToEnd:
     timeout: 5000
 
-# Log out first if a previous session is still active
+# Sign out if a previous session is still active
 - tapOn:
-    id: 'logout-button'
+    id: 'drawer-sign-out'
     optional: true
 - waitForAnimationToEnd:
     timeout: 2000
 - tapOn:
-    text: 'Logout'
+    text: 'Sign Out'
+    optional: true
+- waitForAnimationToEnd:
+    timeout: 3000
+
+# Re-open drawer → Profile
+- tapOn:
+    id: 'profile-icon'
     optional: true
 - waitForAnimationToEnd:
     timeout: 3000
 - tapOn:
-    id: 'profile-icon'
-    optional: true
+    id: 'drawer-row-profile'
 - waitForAnimationToEnd:
     timeout: 3000
 
@@ -57,7 +63,7 @@ name: Nostr Login Test (dev variant)
 - waitForAnimationToEnd:
     timeout: 10000
 
-# Sheet dismisses on success; we should be back on Account
+# Sheet dismisses on success; we should be back on the Profile sub-screen
 - extendedWaitUntil:
     visible:
       text: 'Ben'
