+++ tests/e2e/test-login-perf.yaml
@@ -15,29 +15,32 @@ name: Login to Friends Performance Test
 - waitForAnimationToEnd:
     timeout: 3000
 
-# Navigate to Account via the profile icon
+# Open drawer
 - tapOn:
     id: 'profile-icon'
 - waitForAnimationToEnd:
     timeout: 5000
 
-# Logout if already logged in
+# Sign out if already logged in
 - tapOn:
-    id: 'logout-button'
+    id: 'drawer-sign-out'
     optional: true
 - waitForAnimationToEnd:
     timeout: 2000
-# Confirm logout dialog
 - tapOn:
-    text: 'Logout'
+    text: 'Sign Out'
     optional: true
 - waitForAnimationToEnd:
     timeout: 3000
 
-# Re-navigate to Account (logout may have navigated away)
+# Re-open drawer → Profile
 - tapOn:
     id: 'profile-icon'
     optional: true
+- waitForAnimationToEnd:
+    timeout: 3000
+- tapOn:
+    id: 'drawer-row-profile'
 - waitForAnimationToEnd:
     timeout: 3000
